ajustes controllers e services

+++ src/categoria/services/categoria.service.ts
@@ -37,16 +37,59 @@ export class CategoriaService {
   }
 
   async create(categoria: Categoria): Promise<Categoria> {
+    // Verificar se já existe categoria com mesmo nome
+    const categoriaExistente = await this.categoriaRepository.findOne({
+      where: { nome: categoria.nome },
+    });
+
+    if (categoriaExistente)
+      throw new HttpException(
+        'Categoria com este nome já existe!',
+        HttpStatus.BAD_REQUEST,
+      );
+
     return await this.categoriaRepository.save(categoria);
   }
 
   async update(categoria: Categoria): Promise<Categoria> {
-    await this.findById(categoria.id);
+    await this.findById(categoria.id); // Verifica se a categoria existe
+
+    const categoriaExistente = await this.categoriaRepository.findOne({
+      where: {
+        nome: categoria.nome,
+      },
+    });
+
+    if (categoriaExistente && categoriaExistente.id !== categoria.id)
+      throw new HttpException(
+        'Categoria com este nome já existe!',
+        HttpStatus.BAD_REQUEST,
+      );
+
     return await this.categoriaRepository.save(categoria);
   }
 
+  a;
   async delete(id: number): Promise<DeleteResult> {
-    await this.findById(id);
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const categoria = await this.findById(id); // Já valida se existe
+
+    const categoriaComProdutos = await this.categoriaRepository.findOne({
+      where: { id },
+      relations: ['produtos'],
+    });
+
+    if (
+      categoriaComProdutos &&
+      categoriaComProdutos.produtos &&
+      categoriaComProdutos.produtos.length > 0
+    ) {
+      throw new HttpException(
+        'Não é possível excluir categoria com produtos vinculados!',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     return await this.categoriaRepository.delete(id);
   }
 }
